ops/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.6
Deprecated==1.2.14
dune_client==1.7.3
frozenlist==1.4.1
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-secret-manager==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.0
grpcio-status==1.62.0
idna==3.6
mailchimp_marketing @ git+https://github.com/mailchimp/mailchimp-marketing-python.git@3305fa45b3f436767a539c5fba9cb2b0a083d761
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
ndjson==0.3.1
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
types-Deprecated==1.2.9.20240311
types-python-dateutil==2.9.0.20240316
types-PyYAML==6.0.12.20240311
types-requests==2.31.0.20240406
types-setuptools==69.5.0.20240519
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
